requirements.txt
gradio>=4.44.0
python-dotenv>=1.0.1
openai>=1.60.0
azure-identity>=1.17.1
